edit campaign -- edit fish result image

+++ WildFishingNote/src/com/simple/wildfishingnote/utils/FileUtil.java
@@ -36,8 +36,10 @@ public class FileUtil {
         return fileName;
     }
     
-    public static void deleteFile(String fileName) {
-        File file = new File(fileName);
-        file.delete();
-    }
+	public static void deleteFile(String fileName) {
+		File file = new File(fileName);
+		if (file.exists()) {
+			file.delete();
+		}
+	}
 }
